% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_div_single_backward_selection.R
\name{backward_selection_mindiv}
\alias{backward_selection_mindiv}
\title{Backward Selection for Minimum Divergence Schaake Shuffle}
\usage{
backward_selection_mindiv(X_mat, H_mat, backward_num)
}
\arguments{
\item{X_mat}{a matrix where rows correspond to different locations and lead times
for a given ensemble forecast. The columns correspond to the number of members.}

\item{H_mat}{a matrix of historical observations. The rows and columns of this
matrix are arranged in such a way that the observational trajectories are
consistent with the locations and lead times of the rows in \code{X_mat}.}

\item{backward_num}{the number of dates to retain after backward selection}
}
\value{
A reduced version of the matrix for \code{H_mat}
with \code{backward_num} columns.
}
\description{
Does stuff
}
\details{
details about stuff
}
\references{
Scheuerer, Michael, et al.
"A method for preferential selection of dates in the
Schaake shuffle approach to constructing spatiotemporal
forecast fields of temperature and precipitation."
Water Resources Research 53.4 (2017): 3029-3046.
}
\seealso{
\code{\link{get_mindiv_epsilon}}
}
\author{
Kate Saunders and Kirien Whan
}
