% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_member_names.R
\name{create_member_names}
\alias{create_member_names}
\title{Creates member names}
\usage{
create_member_names(prefix_string, num_members, width)
}
\arguments{
\item{num_members}{integer giving the number of ensemble members}

\item{width}{(optional) this is the width for zero padding with the prefix string
(default  is 3)}

\item{M}{prefix string}
}
\value{
a vector containing the names of the ensemble members
}
\description{
Takes a prefix string, ie "fc" and creates member reference names, ie, "fc001" "fc002"
}
\examples{
create_member_names(prefix_string = "fc", num_members = 10)
create_member_names(prefix_string = "fc", num_members = 10, width = 2)

}
\author{
Kate Saunders
}
