% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecc_quantiles.R
\name{get_ecc_quantiles}
\alias{get_ecc_quantiles}
\title{Creates a vector of quantiles}
\usage{
get_ecc_quantiles(m, ecc_type)
}
\arguments{
\item{m}{number of members in the ensemble (must be an integer)}

\item{ecc_type}{one of the characters \code{'R'}, \code{'Q'}, \code{'Q1'},
or \code{'S'}. This character corresponds to the desired ECC sampling method.
There are two types of equi-spaced quantiles, 'Q' and 'Q1'.}
}
\value{
a vector of sample quantiles
}
\description{
This function produces qunatiles for ECC-R (random qunatiles),
ECC-Q (uniform  quantiles) or ECC-S (jittered quantiles).
}
\details{
If the \code{ecc_type} is \code{'R'} then quantiles are randomly sampled.
If the \code{ecc_type} is \code{'Q'} then qunatiles are equally sampled.
#' If the \code{ecc_type} is \code{'Q1'} then quantiles are equally sampled (Bröck et al. 2012)
If the \code{ecc_type} is \code{'S'} then the quantiles are jittered.
}
\examples{
get_ecc_quantiles(3, "R")
get_ecc_quantiles(4, "Q")
get_ecc_quantiles(5, "S")

}
\references{
Schefzik, Roman, Thordis L. Thorarinsdottir, and Tilmann Gneiting.
"Uncertainty quantification in complex simulation models using ensemble
copula coupling." Statistical science 28.4 (2013): 616-640.

Hu, Yiming, et al. "A stratified sampling approach for improved sampling
from a calibrated ensemble forecast distribution." Journal of Hydrometeorology
17.9 (2016): 2405-2417.
}
\seealso{
\code{\link{sample_ecc_members}}
}
\author{
Kate Saunders and Kirien Whan
}
