% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_min_div_epsilon.R
\name{get_mindiv_epsilon}
\alias{get_mindiv_epsilon}
\title{Minimum Divergence Helper}
\usage{
get_mindiv_epsilon(x_vec, y_val, sgn)
}
\arguments{
\item{x_vec}{a vector of giving values of an ensemble forecasts.
The length of the vector corresponds
to the number of ensemble members.}

\item{y_val}{a numeric value of an historical observation}

\item{sgn}{is one of the characters, "+" or "-".}
}
\value{
a value for \eqn{\epsilon_+} (or \eqn{\epsilon_-}).
}
\description{
This function is a helper that computes terms needed to
estimate the minimum divergence between an ensemble forecast
and a set of historical observations.
}
\details{
This function calculates the difference between an
historical observation and the members of an ensemble forecast.
This difference is then passed to a sign function. The ensemble average
of the evaluated sign function is returned.

The sign functions is mathematically written as \eqn{f(x) = (x)_+}. This
is equivalent operation to \eqn{f(x) = max(x, 0)}.

Let $M$ be the number of members, then if the \code{sgn} is "+"
\eqn{ \epsilon_+ = \frac{1}{M} \Sum_{i = 1}^M (y - x_i)_+ },
otherwise, if \code{sgn} is "-"
\eqn{ \epsilon_- = \frac{1}{M} \Sum_{i = 1}^M (x_i - y)_+ },

These epsilon terms are further described in Scheuerer et al. (2017).
They are computed separately for computational efficiency.
}
\examples{
num_members = 10
x = dnorm(seq(-3,3, length.out = num_members))
get_mindiv_epsilon(x_vec = x, y_val = -3, sgn = "+")
get_mindiv_epsilon(x_vec = x, y_val = -3, sgn = "-")
get_mindiv_epsilon(x_vec = x, y_val = 3, sgn = "-")
get_mindiv_epsilon(x_vec = x, y_val = 3, sgn = "-")

}
\references{
Scheuerer, Michael, et al.
"A method for preferential selection of dates in the
Schaake shuffle approach to constructing spatiotemporal
forecast fields of temperature and precipitation."
Water Resources Research 53.4 (2017): 3029-3046.
}
\seealso{
\code{\link{get_minimum_divergence}}
}
\author{
Kate Saunders and Kirien Whan
}
