% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quantiles.R
\name{get_quantiles}
\alias{get_quantiles}
\title{Quantile sampling}
\usage{
get_quantiles(n_members, method, n_reps = 1)
}
\arguments{
\item{n_members}{number of members in the ensemble. This must be an integer.}

\item{method}{character string corresponding to the desired quantile sampling method
and must be one of \code{'R'}, \code{'Q'}, \code{'QJ'} or \code{'QH'}.}

\item{n_reps}{number of replicates of the function corresponding to the
dimension of the forecast}
}
\value{
a matrix of sample quantiles where the number of rows corresponds to
n_reps and the number of columns corresponds to n_members
}
\description{
Creates a matrix of sampled quantiles for drawing new ensemble members
}
\details{
If the \code{method} is \code{'R'} then the quantiles are randomly sampled.
If the \code{method} is \code{'Q'} then the sampled quantiles are equally spaced.
If the \code{method} is \code{'QH'} then the sampled quantiles are equally spaced using a different sampling scheme.
If the \code{method} is \code{'QJ'} then the sampled quantiles are equally spaced and jittered.
}
\examples{

# Examples of the different sampling methods
get_quantiles(n_members = 3, method = "R")
get_quantiles(n_members = 4, method = "Q")
get_quantiles(n_members = 4, method = "QJ")
get_quantiles(n_members = 5, method = "QH")

# Examples showing difference in methods for repeat sampling
get_quantiles(n_members = 3, method = "R", n_reps = 2)
get_quantiles(n_members = 3, method = "Q", n_reps = 2)

}
\seealso{
\code{\link{run_shuffle_template}}
}
\author{
Kate Saunders and Kirien Whan
}
