% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_min_div_total.R
\name{get_total_divergence}
\alias{get_total_divergence}
\title{Total Minimum Divergence}
\usage{
get_total_divergence(X_mat, H_mat)
}
\arguments{
\item{X_mat}{a matrix where rows are different ensemble forecasts and the columns
are the number of members}

\item{H_mat}{a matrix of historical observations.}
}
\value{
A value for the total divergence between a set of ensemble forecasts and
a set of possible dates that will be used in to perform the Schaake shuffle.
}
\description{
This function commutes the
minimum divergence metric between a set of ensemble forecast and a corresponding
set of historical observations.
}
\details{
This function calculates the minimum divergence between a set of
ensemble forecasts and a corresponding set of historical observations, whose
trajectories could be used to create dependence template.

Let the forecast distribution for a fixed location, $s$, and lead time, $t$,
be given by $F_{st}^{\mathcal{f}}(x)$. For a set of historical dates,
$\mathcal{H}$, let the empirical distribution function that corresponds to the
observational trajectories of the dates in $\mathcal{H}$ be
given by $F_{st}^{\mathcal{H}}(x)$. The minimum divergence is
$\Delta_{st}^{\mathcal{H}} = \int \left(F_{st}^{\mathcal{H}}(x) - F_{st}^{\mathcal{f}}(x) \right)^2 \text{dx}
The total divergence is then given by summing the diverage for all locations and lead times.

This function operates on the matrices row-wise, so a row of $X_mat$ corresponds
to one fixed $s$ and $t$. The observational entries in the $H_mat$ need to be
consistent with these $s$ and $t$.
}
\examples{
num_members = 50
x = dnorm(seq(-3,3, length.out = num_members))
print("FINISH EXAMPLE")

}
\references{
Scheuerer, Michael, et al.
"A method for preferential selection of dates in the
Schaake shuffle approach to constructing spatiotemporal
forecast fields of temperature and precipitation."
Water Resources Research 53.4 (2017): 3029-3046.
}
\seealso{
\code{\link{get_mindiv_epsilon}}
}
\author{
Kate Saunders and Kirien Whan
}
