% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_missing_values.R
\name{interpolate_missing_values}
\alias{interpolate_missing_values}
\title{Interpolates missing observations}
\usage{
interpolate_missing_values(M)
}
\arguments{
\item{M}{is a matrix where columns corresponds to different sample observations}
}
\value{
M with any missing values interpolated
}
\description{
Occasionally missing observations are present in climatologically sample used in the
Schaake shuffle. If only a few of observations are missing if can be useful to interpolate
the missing values.
}
\details{
For each column with a missing observation, the next closest columns is found
in terms of the mean square error. If this column has a valid observation in the same
row as the missing observation, then this value is used to interpolate the missing
observation. This is a simple method of data imputation. Care should be taken that
the reason the observation was missing was not for a systematic reason.
}
\examples{

M = matrix(c(1,2,3,1,NA,3,4,5,6), nrow = 3)
interpolate_missing_values(M)

M = matrix(c(-1,-2,-3,1,NA,3,4,5,6), nrow = 3)
interpolate_missing_values(M)

}
\author{
Kate Saunders
}
