% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_members.R
\name{reorder_members}
\alias{reorder_members}
\title{Reorders ensemble members}
\usage{
reorder_members(X, B)
}
\arguments{
\item{X}{is a matrix where the columns correspond to multivariate forecasts.}

\item{B}{is a matrix with common dimension to X, and contains order statsitics for
reshuffling}
}
\value{
a reshuffled version of matrix according the order statistics given in B.
}
\description{
Reorders the row entries of a matrix using a template based on order statistics
}
\details{
Each columns of X corresponds to an ensemble member. The order statistics in B are
generated from climatologically similar days to the forecast day.
This function is internal and used within \code{schaake_shuffle()}
}
\examples{

X = matrix(c(2,1,3, 5,6,7), nrow = 2, byrow = TRUE)
B = matrix(c(2,1,3, 3,2,1), nrow = 2, byrow = TRUE)
reorder_members(X, B)

}
\author{
Kate Saunders and Kirien Whan
}
