% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shuffle_template.R
\name{run_shuffle_template}
\alias{run_shuffle_template}
\title{Shuffles a post-processed forecast to restore dependence}
\usage{
run_shuffle_template(forecast, template)
}
\arguments{
\item{forecast}{is a matrix where the columns correspond to members with the unrealistic dependence structure, and the rows are a space or time dimension}

\item{template}{is a matrix where the columns correspond to the template for dependence structure, and the rows are a space or time dimension}
}
\value{
a matrix where the post-processed forecast with dependence based on
the template
}
\description{
Shuffles a post-processed forecast to restore dependence
}
\details{
For a forecast that has been sampled from a post-processed forecast distribution
the members need to be reshuffled to restore a meaningful dependence structure
in space, time or both.

This function uses a template to reshuffle the sample ensemble members. The
template can be one of climatology, as in the Schaake Shuffle, or can be from
the raw ensemble forecast, as in Empirical Copula Coupling.
}
\examples{

d <- 3  # number of dimensions (space x time)
m <- 5  # number of members (sampled from the post-process forecast distribution)

set.seed(1)
eg_forecast = matrix(rnorm(d*m), nrow = d)

eg_template = matrix(rnorm(d*m), nrow = d)

forecast_with_dependence = run_shuffle_template(eg_forecast, eg_template)

# Checking the eg_forecast inherited the order from the dependence eg_template
order(eg_forecast[1,])
order(eg_template[1,])
order(forecast_with_dependence[1,])

# Checking the eg_forecast still contains the same members after shuffling
sort(eg_forecast[1,])
sort(forecast_with_dependence[1,])

Add in the example with the data here

}
\references{
Clark, Martyn, et al. "The Schaake shuffle: A method for reconstructing
space–time variability in forecasted precipitation and temperature fields."
Journal of Hydrometeorology 5.1 (2004): 243-262.

Schefzik, Roman, Thordis L. Thorarinsdottir, and Tilmann Gneiting.
"Uncertainty quantification in complex simulation models using ensemble
copula coupling." Statistical science 28.4 (2013): 616-640.

Hu, Yiming, et al. "A stratified sampling approach for improved sampling
from a calibrated ensemble forecast distribution." Journal of Hydrometeorology
17.9 (2016): 2405-2417.
}
\author{
Kate Saunders and Kirien Whan
}
