% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ecc_members.R
\name{sample_ecc_members}
\alias{sample_ecc_members}
\title{Samples ensemble members from a normal distribution}
\usage{
sample_ecc_members(num_members, function_type, pars, ecc_type)
}
\arguments{
\item{num_members}{number of members in the ensemble (must be an integer)}

\item{function_type}{a function to simulate the members from}

\item{pars}{a data frame with named columns corresponding to parameters}

\item{ecc_type}{one of the characters, \code{'R'}, \code{'Q'} or \code{'S'}.
This character corresponds to the desired ECC sampling method}
}
\value{
a matrix where the columns give the sampled ensemble members
}
\description{
This functions samples ensembe members from the provided distribution function.
The type of sampling can be one of ECC-R (random),
ECC-Q (uniform  quantiles) or ECC-S (jittered quantiles).
}
\details{
If the \code{ecc_type} is \code{'R'} then the function_type should
be for random sampling of quantiles, ie. \code{`rnorm`}. If the
\code{ecc_type} is \code{'Q'} or \code{'S'} then the function_type should
be for quantile sampling, ie \code{`qnorm`}.

This function uses \code{\link{get_ecc_quantiles}} for quantile sampling.
}
\examples{

num_members = 5
pars <- data.frame(mean = 0, sd = 1)
sample_ecc_members(num_members,  rnorm, pars, 'R')
sample_ecc_members(num_members, qnorm,  pars,  'Q')
sample_ecc_members(num_members, qnorm,  pars, 'S')

pars <- data.frame(mean = c(0,10), sd = c(1,1))
sample_ecc_members(num_members,  rnorm, pars, 'R')

num_members = 4
pars <- data.frame(rate = c(1,2,3))
sample_ecc_members(num_members, rexp,  pars, 'R')

}
\references{
Schefzik, Roman, Thordis L. Thorarinsdottir, and Tilmann Gneiting.
"Uncertainty quantification in complex simulation models using ensemble
copula coupling." Statistical science 28.4 (2013): 616-640.

Hu, Yiming, et al. "A stratified sampling approach for improved sampling
from a calibrated ensemble forecast distribution." Journal of Hydrometeorology
17.9 (2016): 2405-2417.
}
\seealso{
\code{\link{get_ecc_quantiles}} and \code{\link{reorder_members}}
}
\author{
Kate Saunders and Kirien Whan
}
