% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ecc_members_norm.R
\name{sample_ecc_members_norm}
\alias{sample_ecc_members_norm}
\title{Samples ensemble members from a normal distribution}
\usage{
sample_ecc_members_norm(num_members, pars, draw_type = "R")
}
\arguments{
\item{num_members}{number of members in the ensemble (must be an integer)}

\item{pars}{a data frame with named columns \code{mu} and \code{sigma}}

\item{draw_type}{one of the characters, \code{'R'}, \code{'Q'} or \code{'S'}.
This character corresponds to the desired ECC sampling method}
}
\value{
a matrix where the rows give the ensemble members sampled using
the paremeters given by the the corresponding row of \code{pars}.
}
\description{
This functions samples ensembe members from a normal distribution using
the parameters provided. The type of sampling can be one of ECC-R (random),
ECC-Q (uniform  quantiles) or ECC-S (random quantiles).
}
\examples{

pars <- data.frame(mu = c(1,2), sigma  = c(1,1))
sample_members_norm(5,  pars,  'R')
sample_members_norm(5,  pars,  'Q')
sample_members_norm(5,  pars,  'S')

}
\references{
Schefzik, Roman, Thordis L. Thorarinsdottir, and Tilmann Gneiting.
"Uncertainty quantification in complex simulation models using ensemble
copula coupling." Statistical science 28.4 (2013): 616-640.

Hu, Yiming, et al. "A stratified sampling approach for improved sampling
from a calibrated ensemble forecast distribution." Journal of Hydrometeorology
17.9 (2016): 2405-2417.
}
\author{
Kate Saunders
}
