% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_schaake_dates.R
\name{sample_schaake_dates}
\alias{sample_schaake_dates}
\title{Samples dates for schaake shuffle}
\usage{
sample_schaake_dates(num_draws, dates, date_val, window = 7)
}
\arguments{
\item{num_draws}{number of dates to samples}

\item{dates}{vector of all possible dates where there are suitable observations. This vector will be cross-referenced with the climate window}

\item{date_val}{date of the observation for which a similar climatology is required}

\item{window}{integer that gives the radius of the date window, \code{date_val} +- \code{window} (unit is days)}
}
\value{
a vector of length \code{num_draws} that gives the sampled dates
}
\description{
For a given date, this function samples dates within a surrounding window across different years.
These dates can be used in the Schaake shuffle to generate a dependence template
that is based on climatology.
}
\details{
This function assumes the window of interest is in given in number of days.
The dates input is for handling dates with missing observations.
}
\examples{
date_val = lubridate::as_date("2019-01-01")
dates = seq(lubridate::as_date("2018-01-01"), lubridate::as_date("2020-01-01"), by = "days")

set.seed(1)
sampled_dates <- sample_schaake_dates(num_draws = 3, dates = dates, date_val = date_val,  window = 7)
sampled_dates

set.seed(1)
new_dates = setdiff(dates, sampled_dates) \%>\% lubridate::as_date()
new_sampled_dates <- sample_schaake_dates(num_draws = 3, dates = new_dates, date_val = date_val, window = 7)
new_sampled_dates

}
\references{
Clark, Martyn, et al. "The Schaake shuffle: A method for reconstructing
space–time variability in forecasted precipitation and temperature fields."
Journal of Hydrometeorology 5.1 (2004): 243-262.
}
\author{
Kate Saunders and Kirien Whan
}
