% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaake_shuffle.R
\name{schaake_shuffle}
\alias{schaake_shuffle}
\title{Performs the schaake shuffle}
\usage{
schaake_shuffle(X, Y)
}
\arguments{
\item{X}{is a matrix where the columns correspond to multivariate forecasts.}

\item{Y}{is a matrix where the columns give climatologically similar observations to
that of the forecast day.}
}
\value{
a matrix where the forecasts in X have been reshuffled according to the
dependence inherited from the climatological template in Y.
}
\description{
For a forecast that has been subject to univariate post-processing the members
need to be reshuffled to so that the forecast has the  correct dependence structure.
This function does the reshuffling using a dependence structure that is inherited from
days that have are climatologically similar.
}
\details{
For the input matrix X, the number of rows correspond to the dimension of the
multivariate forecast and the number of columns corresponds the ensemble members.
The dimension of Y must correspond to X. To get climatologically similar days use
the function \code{sample_schaake_dates()}.

No missing values should be present in Y. Data imputation or date resampling
should be used.
}
\examples{

# code for this example was based on the function vs_sample()
# in the scoringRules package

d <- 10  # number of dimensions
m <- 50  # number of samples from multivariate forecast distribution
mu0 <- rep(0, d)
mu <- rep(1, d)
S0 <- S <- diag(d)
S[S==0] <- 0.1
S0[S0==0] <- 0.2

# generate samples from multivariate normal distributions
obs <- drop(mu0 + rnorm(d) \\\%*\\\% chol(S0))
climate_example <- replicate(m, drop(mu + rnorm(d) \\\%*\\\% chol(S)))

forecast_example <- matrix(mu0 + rnorm(d*m), nrow = d, ncol = d)

schaake_shuffle(X = forecast_example, Y = climate_example)

}
\references{
Clark, Martyn, et al. "The Schaake shuffle: A method for reconstructing
space–time variability in forecasted precipitation and temperature fields."
Journal of Hydrometeorology 5.1 (2004): 243-262.
}
\author{
Kate Saunders
}
