% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_members.R
\name{shuffle_members}
\alias{shuffle_members}
\title{Function to rank, order or sort ensemble members}
\usage{
shuffle_members(M, type, ...)
}
\arguments{
\item{M}{is a matrix where the columns correspond to multivariate forecasts.}

\item{type}{is a string of either 'rank', 'order' or 'sort'.
This operation will is applied to the rows}
}
\value{
a matrix where the ensemble members in X have had the operation type applied
}
\description{
Applies a function to rank, order or sort the ensemble members.
}
\details{
No missing values should be present in X.

This function is internal and used within \code{schaake_shuffle()}
}
\examples{

M = matrix(c(2,3,1, 5,6,7), nrow = 2, byrow = TRUE)
shuffle_members(M, 'rank')
shuffle_members(M, 'order')
shuffle_members(M, 'sort')

}
\author{
Kate Saunders
}
