% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_div_thin_dates.R
\name{thin_dates}
\alias{thin_dates}
\title{Thins dates for Minimum Divergence Schaake shuffling}
\usage{
thin_dates(Pred_mat, H_mat, retain_N)
}
\arguments{
\item{Pred_mat}{a matrix where rows correspond to different locations and lead times
for a given ensemble forecast. The columns give the lower and upper predictive
intervals for the marginal distribution.}

\item{H_mat}{a matrix of historical observations. The rows and columns of this
matrix are arranged in such a way that the observational trajectories are
consistent with the locations and lead times of the rows in \code{X_mat}.}

\item{retain_N}{the number of dates to retain after thinning}
}
\value{
A reduced version of the matrix for \code{H_mat}
with \code{retain_N} columns.
}
\description{
This function takes a set observational trajectories, comparing these
trajectories with those of the ensemble forecast. Trajectories with
observations falling outside the range of the post-processed ensemble are
eliminated systematically eliminated. Dates are eliminated systematically
in order based on how many observations fall outside the ensemble trajectory,
until the dates are sufficiently thinned.
}
\details{
This function is used to thin the number of observational trajectories
used in minimum divergence schaake shuffling. This initial screening of dates
greatly reduced the computational time needed to estimate the total
divergence and perform backward selection, see Scheuerer et al. (2017).
}
\examples{
#WRITE AN EXAMPLE
}
\references{
Scheuerer, Michael, et al.
"A method for preferential selection of dates in the
Schaake shuffle approach to constructing spatiotemporal
forecast fields of temperature and precipitation."
Water Resources Research 53.4 (2017): 3029-3046.
}
\seealso{
\code{\link{get_mindiv_epsilon}}
}
\author{
Kate Saunders and Kirien Whan
}
