% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_min_div_epsilon.R
\name{get_mindiv_epsilon}
\alias{get_mindiv_epsilon}
\title{Minimum Divergence Helper}
\usage{
get_mindiv_epsilon(x_vec, y_val, sgn)
}
\arguments{
\item{x_vec}{a numeric vector corresponding to the ensemble forecast from
a marginal distribution, so the forecast is for a fixed location, lead time,
initialisation time and meteorlogical variable.
The length of \code{x_vec} therefore is the same as the number of ensemble members.}

\item{y_val}{a numeric value of an observation}

\item{sgn}{is one of the characters, "+" or "-".}
}
\value{
a value for $ \epsilon_+$ (or $\epsilon_-$).
These $\epsilon$ terms are needed to calculate the minimum divergence
needed for preferential date selection in the Schaake shuffle.
}
\description{
This function is a helper that computes one of terms needed to estimate the
minimum divergence between the ensemble forecast and historical dates.
}
\details{
This function calculates the difference between an historical observation and
the members of an ensemble forecast. This difference is then passed to a sign function,
written as $(x)_+$ where is equivalent to $max(x, 0)$ and the average
over all ensemble members is returned. If the \code{sgn} is "+" then
$ \epsilon_+ = \frac{1}{M} \Sum_{i = 1}^M (y - x_i)_+ $,
otherwise if \code{sgn} is "-" then
$ \epsilon_- = \frac{1}{M} \Sum_{i = 1}^M (x_i - y)_+ $,
}
\examples{
num_members = 10
x = dnorm(seq(-3,3, length.out = num_members))
get_mindiv_epsilon(x_vec = x, y_val = -3, sgn = "+")
get_mindiv_epsilon(x_vec = x, y_val = -3, sgn = "-")
get_mindiv_epsilon(x_vec = x, y_val = 3, sgn = "-")
get_mindiv_epsilon(x_vec = x, y_val = 3, sgn = "-")

}
\references{
Scheuerer, Michael, et al.
"A method for preferential selection of dates in the
Schaake shuffle approach to constructing spatiotemporal
forecast fields of temperature and precipitation."
Water Resources Research 53.4 (2017): 3029-3046.
}
\seealso{
This function is called by \code{\link{get_minimum_divergence}}
}
\author{
Kate Saunders and Kirien Whan
}
