% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_min_div_value.R
\name{get_minimum_divergence}
\alias{get_minimum_divergence}
\title{Minimum Divergence}
\usage{
get_minimum_divergence(x_vec, hist_vec)
}
\arguments{
\item{x_vec}{a vector of giving values of an ensemble forecast.
The length of the vector corresponds
to the number of ensemble members.}

\item{hist_vec}{a vector of historical observations.}
}
\value{
A value for the minimum divergence
}
\description{
This function calculates the minimum divergence between an
ensemble forecast and a corresponding set of historical observations.
}
\details{
Let the forecast distribution
be given by $F^{\mathcal{f}}(x)$. For a set of historical dates,
$\mathcal{H}$, let the empirical distribution function that corresponds to the
observations of the dates in $\mathcal{H}$ be
given by $F^{\mathcal{H}}(x)$. The minimum divergence is
$$\Delta^{\mathcal{H}} = \int \left(F^{\mathcal{H}}(x) - F^{\mathcal{f}}(x) \right)^2 \text{dx}$$

Computation of this minimum divergence term is greatly simplified using
Appendix A of Scheuerer et al. (2017).
}
\examples{
num_members = 50
x = dnorm(seq(-3,3, length.out = num_members))
print("FINISH EXAMPLE")

}
\references{
Scheuerer, Michael, et al.
"A method for preferential selection of dates in the
Schaake shuffle approach to constructing spatiotemporal
forecast fields of temperature and precipitation."
Water Resources Research 53.4 (2017): 3029-3046.
}
\seealso{
\code{\link{get_total_divergence}} and \code{\link{get_mindiv_epsilon}}.
}
\author{
Kate Saunders and Kirien Whan
}
