% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_schaake_shuffle_dates.R
\name{get_schaake_shuffle_dates}
\alias{get_schaake_shuffle_dates}
\title{Gets schaake shuffle dates}
\usage{
get_schaake_shuffle_dates(obs_datetime, window, init_times, ...)
}
\arguments{
\item{obs_datetime}{- a vector of dates corresponding to the observations.
These should be in the date form of `"\%Y-\%m-\%d \%H:\%M:\%S"`.}

\item{window}{- the length of the forecast window. This object must be specified
as a lubridate object. For example, `hours(48)` or `days(2)`.}

\item{init_times}{- a vector containing strings with the forecast
initialisation times. For example, `c("00")` or `c("00", "12")`.
These strings are converted to numeric variables internally by the function.}
}
\value{
a vector of dates of the form `"\%Y-\%m-\%d \%H:\%M:\%S"` that
are suitable for use in the Schaake shuffle.
}
\description{
This function looks at the past observations and identifies dates that do not
contain missing data and could be used in the schaake shuffle.
}
\details{
When there is missing data in the observations these dates are not suitable to be used
in the Schaake shuffle template. Individual users
may choose to interpolate the missing values. For simplicity however, this
function identifies any dates and initialisation times that would contain
missing observations, and returns a vector of only suitable dates for
consideration.
}
\references{
Clark, Martyn, et al. "The Schaake shuffle: A method for reconstructing
space–time variability in forecasted precipitation and temperature fields."
Journal of Hydrometeorology 5.1 (2004): 243-262.
}
\author{
Kate Saunders and Kirien Whan
}
