% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_min_div_total.R
\name{get_total_divergence}
\alias{get_total_divergence}
\title{Total Minimum Divergence}
\usage{
get_total_divergence(X_mat, H_mat)
}
\arguments{
\item{X_mat}{a matrix where rows correspond to different locations and lead times
for a given ensemble forecast. The columns correspond to the number of members.}

\item{H_mat}{a matrix of historical observations. The rows and columns of this
matrix are arranged in such a way that the observational trajectories are
consistent with the locations and lead times of the rows in \code{X_mat}.}
}
\value{
a numeric value giving the total divergence
}
\description{
This function computes the
minimum divergence between an ensemble forecast and a corresponding
set of historical observations. The sum of the minimum divergence over all
locations and lead times is then returned.
}
\details{
This function operates on the two input matrices row-wise, calculating
the minimum divergence for a given location and lead time. The sum over all
locations and lead times is then returned.
}
\examples{
num_members = 50
x = dnorm(seq(-3,3, length.out = num_members))
print("FINISH EXAMPLE")

}
\references{
Scheuerer, Michael, et al.
"A method for preferential selection of dates in the
Schaake shuffle approach to constructing spatiotemporal
forecast fields of temperature and precipitation."
Water Resources Research 53.4 (2017): 3029-3046.
}
\seealso{
\code{\link{get_minimum_divergence}}
}
\author{
Kate Saunders and Kirien Whan
}
